Split CLI behavior tests by concern
