+++ tests/story_brief/test_main_behavior.py
@@ -1,0 +1,138 @@
+from __future__ import annotations
+
+import sys
+from pathlib import Path
+
+import pytest
+
+from telegraphy.story_brief import generate_story_brief as story_cli
+
+
+def test_main_print_only_calls_pick_story_fields_with_selected_date(
+    monkeypatch: pytest.MonkeyPatch, capsys: pytest.CaptureFixture[str]
+) -> None:
+    captured_date = None
+
+    def fake_pick_story_fields(
+        _rng: object,
+        *,
+        selected_date: object = None,
+        data: object = None,
+    ) -> dict[str, object]:
+        nonlocal captured_date
+        captured_date = selected_date
+        return {"title": "Sample", "word_count_target": 900}
+
+    monkeypatch.setattr(sys, "argv", ["story-brief", "--date", "2001-02-03", "--print-only"])
+    monkeypatch.setattr(story_cli, "pick_story_fields", fake_pick_story_fields)
+    monkeypatch.setattr(
+        story_cli,
+        "to_markdown",
+        lambda _fields, data=None: "---\nmock\n---",
+    )
+
+    story_cli.main()
+
+    output = capsys.readouterr().out
+    assert output == "---\nmock\n---\n"
+    assert captured_date
+    assert captured_date.isoformat() == "2001-02-03"
+
+
+def test_main_lint_dataset_exits_early_without_generating_output(
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    class _Report:
+        has_errors = False
+
+    called = {"lint": 0, "emit": 0, "pick": 0}
+
+    def fake_lint(_data: object) -> _Report:
+        called["lint"] += 1
+        return _Report()
+
+    monkeypatch.setattr(sys, "argv", ["story-brief", "--lint-dataset"])
+    monkeypatch.setattr(story_cli, "get_data", lambda: {"source": "test"})
+    monkeypatch.setattr(story_cli, "lint_story_data", fake_lint)
+    monkeypatch.setattr(
+        story_cli,
+        "_emit_lint_report",
+        lambda _: called.__setitem__("emit", called["emit"] + 1),
+    )
+    monkeypatch.setattr(
+        story_cli,
+        "pick_story_fields",
+        lambda *_args, **_kwargs: called.__setitem__("pick", called["pick"] + 1),
+    )
+
+    story_cli.main()
+
+    assert called == {"lint": 1, "emit": 1, "pick": 0}
+
+
+def test_main_force_flag_allows_overwrite_for_existing_file(
+    tmp_path: Path, monkeypatch: pytest.MonkeyPatch
+) -> None:
+    output_file = tmp_path / "forced.md"
+    output_file.write_text("old-content", encoding="utf-8")
+
+    monkeypatch.setattr(
+        sys,
+        "argv",
+        [
+            "story-brief",
+            "-o",
+            str(tmp_path),
+            "--filename",
+            "forced.md",
+            "--force",
+        ],
+    )
+    monkeypatch.setattr(
+        story_cli, "pick_story_fields", lambda *_args, **_kwargs: {"title": "Forced"}
+    )
+    monkeypatch.setattr(story_cli, "to_markdown", lambda _fields, data=None: "new-content")
+
+    story_cli.main()
+
+    assert output_file.read_text(encoding="utf-8") == "new-content"
+
+
+def test_main_lint_dataset_with_errors_exits_one(monkeypatch: pytest.MonkeyPatch) -> None:
+    class _Report:
+        has_errors = True
+
+    monkeypatch.setattr(sys, "argv", ["story-brief", "--lint-dataset"])
+    monkeypatch.setattr(story_cli, "_get_data_cached", lambda: {})
+    monkeypatch.setattr(story_cli, "lint_story_data", lambda _data: _Report())
+    monkeypatch.setattr(story_cli, "_emit_lint_report", lambda _report: None)
+
+    with pytest.raises(SystemExit, match="1"):
+        story_cli.main()
+
+
+def test_main_validate_strict_failure_exits_without_traceback(
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    def mock_fail(_data: object) -> None:
+        raise ValueError("Strict validation failed: boom")
+
+    monkeypatch.setattr(sys, "argv", ["story-brief", "--validate-strict", "--print-only"])
+    monkeypatch.setattr(story_cli, "_get_data_cached", lambda: {})
+    monkeypatch.setattr(story_cli, "validate_story_data_strict", mock_fail)
+
+    with pytest.raises(SystemExit, match="Strict validation failed: boom"):
+        story_cli.main()
+
+
+def test_main_pick_story_fields_failure_exits_with_message(
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    def mock_fail(*_args: object, **_kwargs: object) -> None:
+        raise ValueError("Need at least two")
+
+    monkeypatch.setattr(sys, "argv", ["story-brief", "--print-only"])
+    monkeypatch.setattr(story_cli, "pick_story_fields", mock_fail)
+
+    with pytest.raises(SystemExit, match="Need at least two"):
+        story_cli.main()
